requirements.txt
dash==2.18.2
plotly==6.5.2
pandas==2.3.3
PyMySQL==1.1.1
python-dotenv==1.0.1
cryptography==45.0.7
pyinstaller==6.13.0
